feat(DOMSource): have interface for DOMSource for other sources to implement
