+++ package.json
@@ -89,7 +89,7 @@
     "testem": "^1.10.3",
     "ts-node": "^1.2.2",
     "tsify": "^1.0.3",
-    "tslint": "^3.14.0-dev.1",
+    "tslint": "^3.15.1",
     "typescript": "^2.0.9",
     "uglify-js": "^2.7.0",
     "validate-commit-msg": "^2.6.1"
