+++ tslint.json
@@ -14,7 +14,6 @@
     "no-var-keyword": true,
     "no-empty": true,
     "no-unused-expression": true,
-    "no-unused-variable": true,
     "no-use-before-declare": true,
     "no-var-requires": false,
     "no-require-imports": false,
